fix: rename pages and components using PascalCase
