+++ src/pages/notFoundPage.tsx
@@ -1,8 +1,8 @@
-const Notfound = () => {
+const NotFound = () => {
   return (
     <div>
       <h1>Страница 404</h1>
     </div>
   );
 };
-export { Notfound };
+export { NotFound };
